fix: consistently show non-mvar term in hovertext
Previously, the hover text would only show the term
in the hovertext if the pretty-printed text was 'atomic',
which excluded expressions like `1 + 1`.
This made the hovertext less useful for inspecting an underscore (`_`),
since only the type (and not the term) would be displayed
for a complex expression.

Change the behavior to hide the term if it's a single
metavariable, and otherwise display both the term and type.
This ensures that we do not produce unhelpful hovertext
like `?m : SomeType` (we still produce just `SomeType`).

We now display both the term and type in a few other cases -
this makes the output a little more verbose, but also ensures
that users can see resolved metavariables in the term.

Closes #5367