fix: make formatter use current token table
Previously the formatter was using the builtin token table rather that the one in the current environment. This could lead to round-tripping failures for user-defined notations.

For an illustrative example, given the following notation
```lean4
infixl:65 "+'" => Int.add
notation:65 a:65 "+'-" b:66 => Int.add a (id b)
```
then `5 +' -1` would parse as `Int.add 5 (-1)` and incorrectly pretty print as `5+'-1`. However, this parses as `Int.add 5 (id 1)`.