fix: make `Repr` deriving instance handle explicit type parameters
The `Repr` deriving instance was assuming that all type parameters are implicit. However, if the parameter came from a type index that was promoted to be a parameter, then it is explicit. The result was that some explicit constructor arguments were not being represented.

Reported [on Zulip](https://leanprover.zulipchat.com/#narrow/stream/113489-new-members/topic/.23eval.20removes.20indexes.20inductive.20.20object.20descriptions/near/472301282).