feat: add DecidableRel instances for BitVec
We follow the recipe for `Decidable` and create a decidable instance for
any bitvector width.

This was motivated in a case-bash that was used to prove an `sdiv` lemma at width 1.
However, due to file ordering constraints, we cannot we use the `DecidableRel`
machinery to implement the lemma `sdiv_eq_and`.