fix: automatic instance names about types with hygienic names should be hygienic
Macros sometimes create auxiliary types and instances about them, and they rely on the instance name generate to create unique names in that case.

This modifies the automatic name generator to add a fresh macro scope to the generated name if any of the constants in the type of the instance themselves have macro scopes.

Closes #2044