chore: move BitVec.udiv/umod/sdiv/smod after add/sub/mul/lt
Divison proofs are more likely to depend on add/sub/mul proofs than the other
way around. This prepares for https://github.com/leanprover/lean4/pull/5609,
which adds division proofs that depend on negation to already be defined.

This PR only moves proofs, but does not change the proofs with the exception
of the last `simp` on toNat_smod being turned into a `simp only` to ensure
the proof stability there.