fix: now linters in general do not run on `#guard_msgs` itself
The `#guard_msgs` command runs the command it is attached to as if it were a top-level command. This is because the top-level command elaborator runs linters, and we are interested in capturing linter warnings using `#guard_msgs`. However, the linters will run on `#guard_msgs` itself, leading sometimes to duplicate warnings (like for the unused variable linter).

Rather than special-casing `#guard_msgs` in every affected linter, this PR special-cases it in the top-level command elaborator itself. Now linters are only run if the command doesn't contain `#guard_msgs`. This way, the linters are only run on the sub-command that `#guard_msgs` runs.

It would probably be best to require that `#guard_msgs` *only* be runnable as a top-level command, since there are ways to suppress linting with the current setup.