fix: when pretty printing constant names, do not use aliases from "non-API" `export`s
Heuristic: an "API" `export` is one that exports a declaration to a parent namespace. This PR makes it so that aliases created by `export` from exporting a name into some parallel namespace are not considered.

This is motivated by Lake's use of `export` to make some declarations like `Lean.Name` permanently available in the `Lake` namespace. A consequence to this is that in docgen, we see `Lean.Name` pretty printed as `Lake.Name`.

Another solution would be to create a `scoped open` command that auto-opens some names whenever a namespace is activated. If that is ever implemented, it would be easy to revert the changes in this PR.

Closes #2524