feat: push/pop tactic API
Adds `pushGoal`/`pushGoals` and `popGoal` for manipulating the goal state. These are an alternative to `replaceMainGoal` and `getMainGoal`, but without needing to reason about goal lists that still contain unassigned metavariables.

Modifies `closeMainGoalUsing`. Now the callback is run in a context where the main goal is not present in the goal list. Furthermore, the `checkUnassigned` argument has been replaced with `checkNewUnassigned`, which checks whether the value assigned to the goal has any *new* metavariables, relative to the start of execution of the callback. This is sufficient for the `exact` tactic.