doc: remove docstring from implicitDefEqProofs (#5751)
this option was added in fb97275dcbb683efe6da87ed10a3f0cd064b88fd to
prepare for #4595, due to boostrapping issues, but #4595 has not landed
yet. This is be very confusing when people discover this option and try
to use it (as I did).

So let's clearly mark this as not yet implemented on `master`, and add
the
docstring only with #4595.