fix: make prefix `!` require no whitespace after it
The high precedence of `!` is causing issues with factorial notation in mathlib.

For example, in the following, the argument to `use` is being parsed as `3 (! contradiction)`.
```
example (h : False) : ∃ n : ℕ, n ≠ n := by
  use 3 !
  contradiction
```

Breaking change: expressions such as `! b` now need to be written as `!b`.

Alternative solution (#5824): we could reduce the precedence of prefix `!` to 'lead'.

[Zulip discussion](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/precedence.20of.20Nat.2Efactorial/near/478253033)