[sharecommon] Only consider salient bytes in eq, hash
Change "lean_sharecommon_{eq,hash}" to only consider the salient bytes
of an object, and not any bytes of any unspecified/uninitialized unused
capacity.

Accessing uninitialized storage results in undefined behaviour.

This does not seem to have any semantics disadvantages: If objects
compare equal after this change, their salient bytes are still
equal. By contrast, if the actual identity of allocations needs to be
distinguished, that can be done by just comparing pointers to the
storage.

If we wanted to retain the current logic, we would need initialize the
otherwise unused parts to some specific value to avoid the undefined
behaviour.